requirements.txt
fastapi>=0.115.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
